+++ src/main/java/embl/ebi/variation/eva/pipeline/steps/decider/OptionalDecider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package embl.ebi.variation.eva.pipeline;
+package embl.ebi.variation.eva.pipeline.steps.decider;
 
 import org.opencb.datastore.core.ObjectMap;
 import org.slf4j.Logger;
