+++ src/main/java/embl/ebi/variation/eva/pipeline/steps/tasklet/VariantsLoad.java
@@ -13,30 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package embl.ebi.variation.eva.pipeline.steps;
+package embl.ebi.variation.eva.pipeline.steps.tasklet;
 
 import embl.ebi.variation.eva.VariantJobsArgs;
 import embl.ebi.variation.eva.utils.URLHelper;
-import org.opencb.biodata.models.variant.VariantSource;
 import org.opencb.datastore.core.ObjectMap;
-import org.opencb.datastore.core.QueryOptions;
 import org.opencb.opencga.storage.core.StorageManagerFactory;
 import org.opencb.opencga.storage.core.variant.VariantStorageManager;
-import org.opencb.opencga.storage.core.variant.adaptors.VariantDBAdaptor;
-import org.opencb.opencga.storage.core.variant.stats.VariantStatisticsManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.batch.core.StepContribution;
-import org.springframework.batch.core.configuration.annotation.StepScope;
+import org.springframework.batch.core.*;
+import org.springframework.batch.core.configuration.annotation.JobScope;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Import;
 import org.springframework.stereotype.Component;
 
 import java.net.URI;
-import java.net.URISyntaxException;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 /**
@@ -45,11 +40,10 @@ import java.nio.file.Paths;
  * @author Jose Miguel Mut Lopez &lt;jmmut@ebi.ac.uk&gt;
  */
 @Component
-@StepScope
-@Import({VariantJobsArgs.class})
-public class VariantsStatsLoad implements Tasklet {
-    private static final Logger logger = LoggerFactory.getLogger(VariantsStatsLoad.class);
-    public static final String SKIP_STATS_LOAD = "statistics.load.skip";
+@JobScope
+public class VariantsLoad implements Tasklet {
+    private static final Logger logger = LoggerFactory.getLogger(VariantsLoad.class);
+    public static final String SKIP_LOAD = "load.skip";
 
     @Autowired
     private VariantJobsArgs variantJobsArgs;
@@ -59,21 +53,30 @@ public class VariantsStatsLoad implements Tasklet {
         ObjectMap variantOptions = variantJobsArgs.getVariantOptions();
         ObjectMap pipelineOptions = variantJobsArgs.getPipelineOptions();
 
-        if (pipelineOptions.getBoolean(SKIP_STATS_LOAD)) {
-            logger.info("skipping stats loading");
+        if (pipelineOptions.getBoolean(SKIP_LOAD)) {
+            logger.info("skipping load step, skipLoad is set to {}", pipelineOptions.getBoolean(SKIP_LOAD));
         } else {
-            VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();
-            QueryOptions statsOptions = new QueryOptions(variantOptions);
-            VariantStatisticsManager variantStatisticsManager = new VariantStatisticsManager();
-            VariantDBAdaptor dbAdaptor = variantStorageManager.getDBAdaptor(variantOptions.getString("dbName"), variantOptions);
+            VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();// TODO add mongo
             URI outdirUri = URLHelper.createUri(pipelineOptions.getString("output.dir"));
-            VariantSource variantSource = variantOptions.get(VariantStorageManager.VARIANT_SOURCE, VariantSource.class);
-            URI statsOutputUri = outdirUri.resolve(VariantStorageManager.buildFilename(variantSource));
+            URI nextFileUri = URLHelper.createUri(pipelineOptions.getString("input.vcf"));
 
-            // actual stats load
-            variantStatisticsManager.loadStats(dbAdaptor, statsOutputUri, statsOptions);
+//          URI pedigreeUri = pipelineOptions.getString("input.pedigree") != null ? createUri(pipelineOptions.getString("input.pedigree")) : null;
+            Path output = Paths.get(outdirUri.getPath());
+            Path input = Paths.get(nextFileUri.getPath());
+            Path outputVariantJsonFile = output.resolve(input.getFileName().toString() + ".variants.json" + pipelineOptions.getString("compressExtension"));
+//          outputFileJsonFile = output.resolve(input.getFileName().toString() + ".file.json" + config.compressExtension);
+            URI transformedVariantsUri = outdirUri.resolve(outputVariantJsonFile.getFileName().toString());
+
+
+            logger.info("-- PreLoad variants -- {}", nextFileUri);
+            variantStorageManager.preLoad(transformedVariantsUri, outdirUri, variantOptions);
+            logger.info("-- Load variants -- {}", nextFileUri);
+            variantStorageManager.load(transformedVariantsUri, variantOptions);
+//          logger.info("-- PostLoad variants -- {}", nextFileUri);
+//          variantStorageManager.postLoad(transformedVariantsUri, outdirUri, variantOptions);
         }
 
         return RepeatStatus.FINISHED;
     }
+
 }
